pyproject.toml
[project]
name = "heidi"
version = "0.1.0"
description = "Clinical Decision Support System with multi-server MCP architecture"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43.1",
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.121.2",
    "fastmcp>=2.13.0.2",
    "httpx>=0.27.0",
    "lxml>=6.0.2",
    "mcp>=1.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "streamlit>=1.30.0",
    "uvicorn>=0.38.0",
]
